+++ sdk/cosmos/azure-mgmt-cosmosdb/azure/mgmt/cosmosdb/models/_cosmos_db_management_client_enums.py
@@ -171,7 +171,25 @@ class PrimaryAggregationType(str, Enum):
     last = "Last"
 
 
+class ApiType(str, Enum):
+
+    mongo_db = "MongoDB"
+    gremlin = "Gremlin"
+    cassandra = "Cassandra"
+    table = "Table"
+    sql = "Sql"
+    gremlin_v2 = "GremlinV2"
+
+
 class RoleDefinitionType(str, Enum):
 
     built_in_role = "BuiltInRole"
     custom_role = "CustomRole"
+
+
+class OperationType(str, Enum):
+
+    create = "Create"
+    replace = "Replace"
+    delete = "Delete"
+    system_operation = "SystemOperation"